fixed Rebase Conflicts
